Add functionality to pgbackman_maintenance to stop automatically all backup definitions for databases that have been deleted with DROP DATABASE or renamed in the PgSQL nodes running them.
The affected backup definitions will get a DELETED status and all
files associated to the affected backup definitions will be kept only
for the period of time defined by the automatic_deletion_retention
parameter for the PgSQL node that was running the deleted database.

The associated files will get deleted regardless of the retention_period
and retention_redundancy values for the affected backup definitions.